split wallet into "lib" and "main" modules
